pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "counselai"
version = "0.1.0"
description = "Evidence-based student counselling platform"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy[asyncio]>=2.0.25",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.9",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "dramatiq[redis]>=1.16.0",
    "librosa>=0.10.0",
    "parselmouth>=0.4.3",
    "opencv-python-headless>=4.8.0",
    "numpy>=1.24.0",
    "openai>=1.0.0",
    "google-genai>=1.0.0",
    "httpx>=0.27.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "playwright>=1.41.0",
]

[tool.setuptools.packages.find]
where = ["src"]
